november

+++ index.js
@@ -13,9 +13,9 @@ const serviceAccount = require("./serviceAccountKey.json");
 const { fstat } = require("fs");
 
 admin.initializeApp({
-    credential: admin.credential.cert(serviceAccount),
-    databaseURL: "https://server-auth-41acc.firebaseio.com",
-  });
+  credential: admin.credential.cert(serviceAccount),
+  databaseURL: "https://server-auth-41acc.firebaseio.com",
+});
 const firestore = admin.firestore();
 const csrfMiddleware = csrf({ cookie: true });
 //file upload controller;
@@ -38,9 +38,9 @@ const productStorage = multer.diskStorage({
 
 });
 
-const upload = multer({storage: storage})
+const upload = multer({ storage: storage })
 
-const productUpload = multer({storage: productStorage})
+const productUpload = multer({ storage: productStorage })
 
 const port = process.env.PORT || 3000;
 const app = express();
@@ -54,81 +54,81 @@ app.use(csrfMiddleware);
 
 app.use(function (req, res, next) {
   var socials = JSON.parse(fs.readFileSync('./public/socials.json'));
-    var token = req.csrfToken();
-    res.cookie("XSRF-TOKEN", token);
-    res.locals.csrfToken = token;
-    res.locals.socials = socials;
-    next();
-  });
+  var token = req.csrfToken();
+  res.cookie("XSRF-TOKEN", token);
+  res.locals.csrfToken = token;
+  res.locals.socials = socials;
+  next();
+});
 //public routes
 app.get('/', async (req, res) => {
   var data = JSON.parse(fs.readFileSync('./public/home.json'));
-  res.status(200).render('client/index', {data});
-      
+  res.status(200).render('client/index', { data });
+
 });
 app.get('/company', (req, res) => {
   var data = JSON.parse(fs.readFileSync('./public/company.json'));
   var items = data.items
-  res.render('client/Бизнес', {items})
+  res.render('client/Бизнес', { items })
 })
 app.get('/about', (req, res) => {
   var data = JSON.parse(fs.readFileSync('./public/about.json'));
-  res.status(200).render('client/Бидний-Тухай', {data});
+  res.status(200).render('client/Бидний-Тухай', { data });
 });
 app.get('/sales', async (req, res) => {
   const product = firestore.collection('products').orderBy('timestamp');
-        const data = await product.get();
-        const products = [];
-        if(data.empty) {
-            return null;
-        }else {
-            data.forEach(doc => {
-                const product = new Product(
-                  doc.id,
-                  doc.data().name,
-                  doc.data().desc1,
-                  doc.data().desc2,
-                  doc.data().category,
-                  doc.data().top1,
-                  doc.data().top2,
-                  doc.data().top3,
-                  doc.data().mid,
-                  doc.data().bottomList,
-                  doc.data().midTitle,
-                  doc.data().midTSubtitle,
-                  doc.data().headline,
-                  doc.data().article,
-                  doc.data().quote,
-                );
-                products.push(product);
-            });
-            
-        }
+  const data = await product.get();
+  const products = [];
+  if (data.empty) {
+    return null;
+  } else {
+    data.forEach(doc => {
+      const product = new Product(
+        doc.id,
+        doc.data().name,
+        doc.data().desc1,
+        doc.data().desc2,
+        doc.data().category,
+        doc.data().top1,
+        doc.data().top2,
+        doc.data().top3,
+        doc.data().mid,
+        doc.data().bottomList,
+        doc.data().midTitle,
+        doc.data().midTSubtitle,
+        doc.data().headline,
+        doc.data().article,
+        doc.data().quote,
+      );
+      products.push(product);
+    });
+
+  }
   res.status(200).render('client/Борлуулалт', { products });
 });
 app.get('/sales/:id', async (req, res) => {
-  
+
   const id = req.params.id;
   const ref = firestore.collection('products').doc(id);
-      const data = await ref.get();
-      var product = {};
-      if(!data.exists) {
-          res.status(404)
-      }else {
-          product = data.data();
-          
-      }
-    res.render('client/product', {product , id});
-  
-  
+  const data = await ref.get();
+  var product = {};
+  if (!data.exists) {
+    res.status(404)
+  } else {
+    product = data.data();
+
+  }
+  res.render('client/product', { product, id });
+
+
 });
 app.get('/contact', (req, res) => {
   res.status(200).render('client/Холбоо-барих');
 });
 //admin routes
 app.get("/admin", (req, res) => {
-    const sessionCookie = req.cookies.session || "";
-    admin
+  const sessionCookie = req.cookies.session || "";
+  admin
     .auth()
     .verifySessionCookie(sessionCookie, true /** checkRevoked */)
     .then((userData) => {
@@ -138,50 +138,50 @@ app.get("/admin", (req, res) => {
     .catch((error) => {
       res.redirect("/login");
     });
-    
+
 });
 app.get('/admin/sales', async (req, res) => {
-    const sessionCookie = req.cookies.session || "";
-    const product = firestore.collection('products').orderBy('timestamp');
-        const data = await product.get();
-        const products = [];
-        if(data.empty) {
-            return null;
-        }else {
-            data.forEach(doc => {
-                const product = new Product(
-                    doc.id,
-                    doc.data().name,
-                    doc.data().desc1,
-                    doc.data().desc2,
-                    doc.data().category,
-                    doc.data().top1,
-                    doc.data().top2,
-                    doc.data().top3,
-                    doc.data().mid,
-                    doc.data().bottomList,
-                    doc.data().midTitle,
-                    doc.data().midTSubtitle,
-                    doc.data().headline,
-                    doc.data().article,
-                    doc.data().quote,
-                );
-                products.push(product);
-            });
-            
-        }
-    
-    admin
+  const sessionCookie = req.cookies.session || "";
+  const product = firestore.collection('products').orderBy('timestamp');
+  const data = await product.get();
+  const products = [];
+  if (data.empty) {
+    return null;
+  } else {
+    data.forEach(doc => {
+      const product = new Product(
+        doc.id,
+        doc.data().name,
+        doc.data().desc1,
+        doc.data().desc2,
+        doc.data().category,
+        doc.data().top1,
+        doc.data().top2,
+        doc.data().top3,
+        doc.data().mid,
+        doc.data().bottomList,
+        doc.data().midTitle,
+        doc.data().midTSubtitle,
+        doc.data().headline,
+        doc.data().article,
+        doc.data().quote,
+      );
+      products.push(product);
+    });
+
+  }
+
+  admin
     .auth()
     .verifySessionCookie(sessionCookie, true /** checkRevoked */)
     .then((userData) => {
       console.log("Logged in:", userData.email)
-      res.render('admin/sales', {products});
+      res.render('admin/sales', { products });
     })
     .catch((error) => {
       res.redirect("/login");
     });
-    
+
 });
 app.post('/admin/sales/create', async (req, res) => {
   const data = {
@@ -207,50 +207,50 @@ app.post('/admin/sales/create', async (req, res) => {
     res.set('Content-Type', 'text/html');
     res.status(201).send('Done');
   } catch (e) {
-      res.status(500);
-      console.log(e)
+    res.status(500);
+    console.log(e)
   }
-  
+
   console.log('data created')
 });
 
 app.get('/admin/sales/:id', async (req, res) => {
-    const sessionCookie = req.cookies.session || "";
-    const id = req.params.id;
-    const ref = firestore.collection('products').doc(id);
-        const data = await ref.get();
-        var product = {};
-        if(!data.exists) {
-            res.status(404)
-        }else {
-            product = data.data();
-            
-        }
-    admin
+  const sessionCookie = req.cookies.session || "";
+  const id = req.params.id;
+  const ref = firestore.collection('products').doc(id);
+  const data = await ref.get();
+  var product = {};
+  if (!data.exists) {
+    res.status(404)
+  } else {
+    product = data.data();
+
+  }
+  admin
     .auth()
     .verifySessionCookie(sessionCookie, true /** checkRevoked */)
     .then((userData) => {
       console.log("Logged in:", userData.email)
-      res.render('admin/product', {product , id});
+      res.render('admin/product', { product, id });
     })
     .catch((error) => {
       res.redirect("/login");
     });
-    
+
 });
 app.post('/admin/sales/delete', async (req, res) => {
   const sessionCookie = req.cookies.session || "";
   const id = req.body.id;
   const ref = firestore.collection('products').doc(id);
   try {
-      await ref.delete();
-      res.set('Content-Type', 'text/html');
-      res.status(201).send('Done');
+    await ref.delete();
+    res.set('Content-Type', 'text/html');
+    res.status(201).send('Done');
   } catch (e) {
-      res.status(500);
-      console.log(e)
+    res.status(500);
+    console.log(e)
   }
-  
+
 });
 app.post('/admin/sales/update-text', async (req, res) => {
   const sessionCookie = req.cookies.session || "";
@@ -267,24 +267,24 @@ app.post('/admin/sales/update-text', async (req, res) => {
   }
 
   admin
-  .auth()
-  .verifySessionCookie(sessionCookie, true /** checkRevoked */)
-  .then(async (userData) => {
+    .auth()
+    .verifySessionCookie(sessionCookie, true /** checkRevoked */)
+    .then(async (userData) => {
       console.log("Logged in:", userData.email)
       try {
-          await ref.update(formData);
-          res.set('Content-Type', 'text/html');
-          res.status(201).send('Done');
+        await ref.update(formData);
+        res.set('Content-Type', 'text/html');
+        res.status(201).send('Done');
       } catch (e) {
-          res.status(500);
-          console.log(e)
+        res.status(500);
+        console.log(e)
       }
-      
-    
-  })
-  .catch((error) => {
-    res.redirect("/login");
-  });
+
+
+    })
+    .catch((error) => {
+      res.redirect("/login");
+    });
 });
 app.post('/admin/sales/product-save1', productUpload.array('image', 3), async (req, res) => {
   const ref = firestore.collection('products').doc(req.body.id);
@@ -293,9 +293,21 @@ app.post('/admin/sales/product-save1', productUpload.array('image', 3), async (r
   const path1 = './public/images/product/' + doc.top1;
   const path2 = './public/images/product/' + doc.top2;
   const path3 = './public/images/product/' + doc.top3;
-  fs.unlinkSync(path1);
-  fs.unlinkSync(path2);
-  fs.unlinkSync(path3);
+  if (doc.top1 == "default-image.jpg") {
+    console.log('skipped unlink')
+  } else {
+    fs.unlinkSync(path1);
+  }
+  if (doc.top2 == "default-image.jpg") {
+    console.log('skipped unlink')
+  } else {
+    fs.unlinkSync(path2);
+  }
+  if (doc.top3 == "default-image.jpg") {
+    console.log('skipped unlink')
+  } else {
+    fs.unlinkSync(path3);
+  }
   const formData = {
     top1: req.files[1].filename,
     top2: req.files[0].filename,
@@ -310,7 +322,12 @@ app.post('/admin/sales/product-save2', productUpload.single('image', 1), async (
   const oldData = await ref.get();
   const doc = oldData.data();
   const path = './public/images/product/' + doc.mid;
-  fs.unlinkSync(path);
+  if (doc.mid == "default-image.jpg") {
+    console.log('skipped unlink')
+  } else {
+    fs.unlinkSync(path);
+  }
+
   const formData = {
     mid: req.file.filename,
   }
@@ -324,9 +341,13 @@ app.post('/admin/sales/product-save3', productUpload.array('images', 12), async 
   const doc = oldData.data();
   doc.bottomList.forEach(item => {
     const path = './public/images/product/' + item;
-    fs.unlinkSync(path);
+    if (item == "default-image.jpg") {
+      console.log('skipped unlink')
+    } else {
+      fs.unlinkSync(path);
+    }
   });
-  const listOfImages = req.files.map(({filename}) => (filename));
+  const listOfImages = req.files.map(({ filename }) => (filename));
   const formData = {
     bottomList: listOfImages,
   }
@@ -336,60 +357,60 @@ app.post('/admin/sales/product-save3', productUpload.array('images', 12), async 
 });
 
 app.get('/admin/service', async (req, res) => {
-    const sessionCookie = req.cookies.session || "";
-    const service = firestore.collection('services');
-        const data = await service.get();
-        const services = [];
-        if(data.empty) {
-            return null;
-        }else {
-            data.forEach(doc => {
-                const service = new Service(
-                    doc.id,
-                    doc.data().title,
-                    doc.data().subtitle,
-                );
-                services.push(service);
-            });
-            
-        }
-    admin
+  const sessionCookie = req.cookies.session || "";
+  const service = firestore.collection('services');
+  const data = await service.get();
+  const services = [];
+  if (data.empty) {
+    return null;
+  } else {
+    data.forEach(doc => {
+      const service = new Service(
+        doc.id,
+        doc.data().title,
+        doc.data().subtitle,
+      );
+      services.push(service);
+    });
+
+  }
+  admin
     .auth()
     .verifySessionCookie(sessionCookie, true /** checkRevoked */)
     .then((userData) => {
       console.log("Logged in:", userData.email)
-      res.render('admin/Нүүр', {services});
+      res.render('admin/Нүүр', { services });
     })
     .catch((error) => {
       res.redirect("/login");
     });
-    
+
 });
 const adminHome = require('./routes/admin-home');
 app.use('/admin', adminHome)
 
 app.get('/admin/showcase/edit', async (req, res) => {
-    const sessionCookie = req.cookies.session || "";
-    const service = firestore.collection('showcase').doc('nEyjlOFanwfyfYjianar');
-        const data = await service.get();
-        var showcase = {};
-        if(!data.exists) {
-            res.status(404)
-        }else {
-            showcase = data.data();
-        }
-        
-    admin
+  const sessionCookie = req.cookies.session || "";
+  const service = firestore.collection('showcase').doc('nEyjlOFanwfyfYjianar');
+  const data = await service.get();
+  var showcase = {};
+  if (!data.exists) {
+    res.status(404)
+  } else {
+    showcase = data.data();
+  }
+
+  admin
     .auth()
     .verifySessionCookie(sessionCookie, true /** checkRevoked */)
     .then((userData) => {
       console.log("Logged in:", userData.email)
-      res.render('admin/Санал-болгох-засах', {showcase});
+      res.render('admin/Санал-болгох-засах', { showcase });
     })
     .catch((error) => {
       res.redirect("/login");
     });
-    
+
 });
 
 
@@ -406,101 +427,101 @@ app.post('/admin/showcase/upload', upload.array('images', 4), async (req, res) =
     fs.unlinkSync(path4);
   });
   const sessionCookie = req.cookies.session || "";
-  
+
   const formData = {
     title: req.body.title,
     subtitle: req.body.subtitle,
-  } 
+  }
   const service = firestore.collection('showcase').doc('nEyjlOFanwfyfYjianar');
   await service.update(formData);
   admin
-  .auth()
-  .verifySessionCookie(sessionCookie, true /** checkRevoked */)
-  .then(async (userData) => {
-    console.log("Logged in:", userData.email);
-    
-        
-    res.redirect('/admin/showcase');
-  })
-  .catch((error) => {
-    res.redirect("/admin/showcase");
-  });
-  
+    .auth()
+    .verifySessionCookie(sessionCookie, true /** checkRevoked */)
+    .then(async (userData) => {
+      console.log("Logged in:", userData.email);
+
+
+      res.redirect('/admin/showcase');
+    })
+    .catch((error) => {
+      res.redirect("/admin/showcase");
+    });
+
 });
 
 
 
 
 app.get("/login", function (req, res) {
-    res.render("admin/login");
-  });
+  res.render("admin/login");
+});
 
 
 
 app.post("/sessionLogin", (req, res) => {
-    const idToken = req.body.idToken.toString();
-  
-    const expiresIn = 60 * 60 * 24 * 7 * 1000;
-  
-    admin
-      .auth()
-      .createSessionCookie(idToken, { expiresIn })
-      .then(
-        (sessionCookie) => {
-          const options = { maxAge: expiresIn, httpOnly: true };
-          res.cookie("session", sessionCookie, options);
-          res.end(JSON.stringify({ status: "success" }));
-        },
-        (error) => {
-          res.status(401).send("UNAUTHORIZED REQUEST!");
-        }
-      );
-  });
-  
-  app.post("/admin/update-service", async (req, res) => {
-    const sessionCookie = req.cookies.session || "";
-    const ref = firestore.collection('services');
-    const data1 = {
-        title: req.body.title1.replace(/['"]+/g, ''),
-        subtitle: req.body.subtitle1.replace(/['"]+/g, ''),
-    }
-    const data2 = {
-        title: req.body.title2.replace(/['"]+/g, ''),
-        subtitle: req.body.subtitle2.replace(/['"]+/g, ''),
-    }
-    const data3 = {
-        title: req.body.title3.replace(/['"]+/g, ''),
-        subtitle: req.body.subtitle3.replace(/['"]+/g, ''),
-    }
+  const idToken = req.body.idToken.toString();
 
-    admin
+  const expiresIn = 60 * 60 * 24 * 7 * 1000;
+
+  admin
+    .auth()
+    .createSessionCookie(idToken, { expiresIn })
+    .then(
+      (sessionCookie) => {
+        const options = { maxAge: expiresIn, httpOnly: true };
+        res.cookie("session", sessionCookie, options);
+        res.end(JSON.stringify({ status: "success" }));
+      },
+      (error) => {
+        res.status(401).send("UNAUTHORIZED REQUEST!");
+      }
+    );
+});
+
+app.post("/admin/update-service", async (req, res) => {
+  const sessionCookie = req.cookies.session || "";
+  const ref = firestore.collection('services');
+  const data1 = {
+    title: req.body.title1.replace(/['"]+/g, ''),
+    subtitle: req.body.subtitle1.replace(/['"]+/g, ''),
+  }
+  const data2 = {
+    title: req.body.title2.replace(/['"]+/g, ''),
+    subtitle: req.body.subtitle2.replace(/['"]+/g, ''),
+  }
+  const data3 = {
+    title: req.body.title3.replace(/['"]+/g, ''),
+    subtitle: req.body.subtitle3.replace(/['"]+/g, ''),
+  }
+
+  admin
     .auth()
     .verifySessionCookie(sessionCookie, true /** checkRevoked */)
     .then(async (userData) => {
-        console.log("Logged in:", userData.email)
-        try {
-            await ref.doc(req.body.id1).update(data1);
-            await ref.doc(req.body.id2).update(data2);
-            await ref.doc(req.body.id3).update(data3);
-            res.set('Content-Type', 'text/html');
-            res.status(201).send('Done');
-        } catch (e) {
-            res.status(500);
-            console.log(e)
-        }
-        
-      
+      console.log("Logged in:", userData.email)
+      try {
+        await ref.doc(req.body.id1).update(data1);
+        await ref.doc(req.body.id2).update(data2);
+        await ref.doc(req.body.id3).update(data3);
+        res.set('Content-Type', 'text/html');
+        res.status(201).send('Done');
+      } catch (e) {
+        res.status(500);
+        console.log(e)
+      }
+
+
     })
     .catch((error) => {
       res.redirect("/login");
     });
-  });
-  
-  app.get("/sessionLogout", (req, res) => {
-    res.clearCookie("session");
-    res.redirect("/login");
-  });
+});
+
+app.get("/sessionLogout", (req, res) => {
+  res.clearCookie("session");
+  res.redirect("/login");
+});
 
 app.listen(port, () => {
-    console.log(`>DEBUG: Server is listening at http://localhost:${port} `);
+  console.log(`>DEBUG: Server is listening at http://localhost:${port} `);
 });
